build(test.yml): controled failiure to test WF. #147

+++ backend/pawtel/hotels/tests/test_services.py
@@ -51,7 +51,7 @@ class HotelServiceTest(TestCase):
         room_types = HotelService.get_all_room_types_of_hotel(self.hotel.id)
         self.assertEqual(len(room_types), 2)
         self.assertEqual(room_types[0].name, "Single")
-        self.assertEqual(room_types[1].name, "Double")
+        self.assertEqual(room_types[1].name, "Single")
 
     def test_get_total_vacancy_for_each_room_type_of_hotel(self):
         vacancies = HotelService.get_total_vacancy_for_each_room_type_of_hotel(
